Add files via upload
Here we had the data file which provided the data about the areas that are affected by Natural phenomena such as hurricanes or fire. There is some information about the environment in the data file. The explanation of the columns is provided in the following part. The goal is to predict the area that might be affected based on the features. You need to prepossess the data and make visualizations and statistical summary reports. The primary approach would be fitting a regression model for predicting the area that will be affected. You might also have a second approach to the problem as a classification problem that concludes if the area will be affected or not.